robbbbba

+++ backend/app.py
@@ -10,7 +10,7 @@ def giveLocations():
     category = request.args.get('category', type=str)
     longitude = request.args.get('longitude', type=float)
     latitude = request.args.get('latitude', type=float)
-    distance = request.args.get('distance', type=int)
+    distance = request.args.get('distance', type=float)
     lista = service.retrieveLocations(category, longitude, latitude, distance)
     for y in lista:
         y['_id'] = str(y['_id'])
